fix(esp32,ui): resolve marquee ghosting, continuous ticker scrolling, exhaustive i18n & automated test suites
- Fix MarqueeEngine ghosting by clearing the matrix framebuffer upon module transition in RotationManager and during MarqueeEngine::activate().
- Overhaul MessageEngine to use continuous delta-time sub-pixel scrolling (60 FPS) with dt stall clamping and rounded matrix coordinates.
- Expand WebUI internationalization to 354 keys across EN, FR, and ES with 100% dictionary parity, covering all 19 engines, 35+ fields, universal options, and system preferences.
- Implement live rotation state sync across DOM elements on deletion and instant screen_title propagation in rotation dropdowns.
- Add automated unit and integration tests:
  * Unity test_message_scroll_math in test_engines.cpp.
  * Unity test_instance_creation_without_variant and test_rotation_entry_sync in test_api.cpp.
  * Comprehensive validation suite in scripts/validate_ui_and_i18n.py.

+++ src/engines/MessageEngine.h
@@ -47,8 +47,8 @@ private:
     unsigned long startTime;
     unsigned long lastUpdate;
     
-    int cursorX;
-    int cursorY;
+    float cursorX;
+    float cursorY;
     int textWidth;
     int textHeight;
     int baselineOffset;
